chore: 🤖 cleanup some
