+++ packages/engine-ui/src/utils/helpers.ts
@@ -54,6 +54,15 @@ const formatCoordinateToFixedDecimalLength = ({
   length: number
 }) => `${lat.toFixed(length)}, ${lon.toFixed(length)}`
 
+export const hexToRGBA = (hex: string, opacity: number = 255) => {
+  hex = hex.replace('#', '')
+  const r = parseInt(hex.substring(0, 2), 16)
+  const g = parseInt(hex.substring(2, 4), 16)
+  const b = parseInt(hex.substring(4, 6), 16)
+
+  return [r, g, b, opacity]
+}
+
 export default {
   findAddress,
   calculateMinTime,
@@ -61,4 +70,5 @@ export default {
   getLastFourChars,
   withoutLastFourChars,
   formatCoordinateToFixedDecimalLength,
+  hexToRGBA,
 }
